projects/ai-video-tool/requirements.txt
gradio
yt-dlp
chat-downloader
google-genai
ffmpeg-python
python-dotenv
pandas
numpy